Optimize best build matches query for a vcs_revision WHERE predicate
The Postgres query optimizer seems to be applying the WHERE predciate, which filters by 'vcs_revision', too late in the process, which results in materializing a ~1M row matches table before filtering it by commitl. This workaround extracts the logic of multiple views into one large query so that the WHERE predicate can be pushed further inside.

Closes #52